Guard against broken or missing time zone
TimeZone.getTimeZone(...) silently returns GMT when Java time zone data is
broken or missing (e.g. third-party Java 7 on Ubuntu 16.04) or when the
identifier is unknown or has a typo. TimeZoneUtil provides a way to get the
exact time zone requested or throws an exception, somewhat similar to
asking for a character encoding that doesn't exist. Added a test at daemon
startup to check for probably missing or broken time zone data and
exit (would confuse crawl windows).